sam-3d-body/requirements.txt
# Python requirements for backend/app
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
mediapipe==0.10.13
opencv-contrib-python==4.13.0.90
rtmlib==0.0.14
onnxruntime==1.23.2
numpy==2.4.1
pandas==2.3.3
scipy==1.17.0
pillow==12.1.0
pyyaml==6.0.3
streamlit==1.53.1
requests==2.32.5
sqlalchemy==2.0.45
pymysql==1.1.2
google-cloud-storage==3.7.0
ffmpeg-python==0.2.0
google-auth==2.45.0
# google-generativeai>=0.3.0
google-genai==1.4.0
python-dotenv==1.2.1
aiohttp==3.13.3
aiofiles==25.1.0
aiomysql==0.2.0
greenlet==3.0.1
ultralytics==8.4.9
roboflow==1.2.13
lap==0.5.12
neo4j==6.1.0
# OpenSim Python API for 3D stress pipeline (IK, GRF, ID). Supports Python 3.10–3.12.
pyopensim==4.5.2.0
bcrypt==5.0.0
ecdsa==0.19.1
passlib==1.7.4
python-jose==3.5.0
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
tiktoken>=0.5.0

# SAM 3D Body (install torch/torchvision separately from pytorch.org if using CUDA)
torch>=2.0.0
torchvision>=0.15.0
pytorch-lightning==2.6.1
omegaconf==2.3.0
yacs==0.1.8
einops==0.8.2
timm==1.0.25
roma==1.5.6
trimesh==4.11.2
pyrender==0.1.45
pyrootutils==1.0.4
braceexpand==0.1.7
termcolor==3.3.0
smplx==0.1.28
tqdm==4.67.3
scikit-learn==1.8.0
# pymomentum-cpu: optional; requires glibc 2.28+. Use conda or requirements-pymomentum.txt on supported systems.
pymomentum-cpu==0.1.108.post0

# SAM 3D Body → IK in single venv (AddBiomechanics kinematics pass)
nimblephysics==0.10.52.1
matplotlib>=3.5.0
